refactor(web): extract ThemeTileGrid from AppearanceSection
